requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
langchain==0.3.1
langgraph==0.2.28
langchain-anthropic==0.2.3
chromadb==0.5.11
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.6
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
